♻️ rfac(custom): step2, 3 프로젝트 선택 로직 이동
